requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
lightning
kaggle
bentoml
transformers
apache-airflow
pillow
torchvision
matplotlib


# PyTorch specific requirements with a custom index URL
--find-links https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio
